requirements.txt
aicrowd-cli==0.1.2
ants==0.0.7
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.3.4
asn1crypto==0.24.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.2
bokeh==2.3.0
Brlapi==0.6.6
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
cloudpickle==1.6.0
cmake==3.18.4.post1
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
contextvars==2.4
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
dask==2021.3.0
dataclasses==0.8
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distributed==2021.3.0
distro==1.5.0
distro-info===0.18ubuntu0.18.04.1
Django==3.2.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
fsspec==0.9.0
future==0.18.2
gevent==21.1.2
gitdb==4.0.7
GitPython==3.1.14
greenlet==1.0.0
HeapDict==1.0.1
httplib2==0.9.2
idna==2.10
imageio==2.9.0
immutables==0.15
importlib-metadata==4.0.1
importlib-resources==3.3.0
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itk==5.1.2
itk-core==5.1.2
itk-filtering==5.1.2
itk-io==5.1.2
itk-numerics==5.1.2
itk-registration==5.1.2
itk-segmentation==5.1.2
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
locket==0.2.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
mando==0.6.4
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mpmath==1.2.1
msgpack==1.0.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
netifaces==0.10.4
networkx==2.5
nibabel==3.2.1
nilearn==0.7.0
notebook==6.2.0
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
opencv-python==4.5.1.48
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
partd==1.1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
powerline-status==2.6
prometheus-client==0.9.0
prompt-toolkit==3.0.11
protobuf==3.0.0
psutil==5.8.0
ptyprocess==0.7.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydicom==2.1.2
Pygments==2.8.1
pygobject==3.26.1
pymacaroons==0.13.0
PyMCubes==0.1.2
PyNaCl==1.1.2
pyparsing==2.4.7
pypi-xmlrpc==2020.12.3
pyRFC3339==1.0
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2021.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.4.1
pyzmq==21.0.1
qtconsole==5.0.2
QtPy==1.9.0
radon==4.3.2
reportlab==3.4.0
requests==2.25.1
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
rich==10.1.0
scikit-build==0.11.1
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.15.0
smmap==4.0.0
sortedcontainers==2.3.0
speedtest-cli==2.1.2
sqlparse==0.4.1
sympy==1.7.1
system-service==0.3
systemd-python==234
tblib==1.7.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.2
toolz==0.11.1
torch==1.7.1
torchvision==0.8.2
tornado==6.1
tqdm==4.60.0
traitlets==4.3.3
typing-extensions==3.10.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.4
usb-creator==0.3.3
virtualenv==20.2.2
wadllib==1.3.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xkit==0.0.0
zict==2.0.0
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.2.0
django-filter